Refactor(frontend): moved themes out of constants
They're not present in their own file, the constants file is starting to
become a little cluttered. Maybe worth moving all the singular view
specific stuff into their own files rather than in the constant file.
Hard to know which gives better visibility.
