+++ frontend/constants.ts
@@ -1,39 +1,6 @@
-import { DefaultTheme, DarkTheme } from "react-native-paper";
 import * as Yup from "yup";
 
 export const API_URL = "http://localhost:5000/api/v1";
-
-export const colors = {
-  ...DefaultTheme.colors,
-  primary: "#1e88e5",
-  accent: "#1e88e5",
-};
-
-export const lightTheme = {
-  ...DefaultTheme,
-  roundness: 2,
-  colors: {
-    ...DefaultTheme.colors,
-    primary: "#1e88e5",
-    accent: "#1e88e5",
-  },
-};
-
-export const darkTheme = {
-  ...DarkTheme,
-  roundness: 2,
-  colors: {
-    ...DarkTheme.colors,
-    primary: "#1e88e5",
-    accent: "#1e88e5",
-  },
-};
-
-export type ThemeType = "Dark" | "Light";
-
-// Location for {@link AsyncStorage} to store theme
-// "Light" or "Dark"
-export const ThemeStorage = "@theme";
 
 /**
  * Login Validation Schema, a Yup Schema for basic validation for Login not nearly as extensive as Register,
